Float bandgap

+++ Globals.h
@@ -34,7 +34,7 @@ public:
     static const std::vector<std::string> kPersistenceDelimiters;
 
     static const float kBaseProb;
-    static const uint8_t kBandGap;
+    static const float kBandGap;
     static const uint8_t kMaxMultiWordSize;
     static const uint8_t kReferenceFrameSize;
     static const uint8_t kMaxCompletionSize;
